Fixed the location of the functions. Attached the file with the bot code.
